# services/wallet-service/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
...
        "title": "Core Wallet API",
        "termsOfService": "https://docs.wallet.example/terms",
        "contact": {
            "name": "Core Wallet Team",
            "email": "hoalh2@hdbank.com.vn"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/finance/merchant-withdraw": {
            "post": {
                "description": "Withdraw from a merchant group settlement account, sweeping hot shards first when `auto_sweep` is true (default).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Merchant settlement withdraw",
                "parameters": [
                    {
                        "description": "Merchant withdraw request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.MerchantWithdrawRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Duplicate idempotent replay",
                        "schema": {
                            "$ref": "#/definitions/dto.MerchantWithdrawResponse"
                        }
                    },
                    "201": {
                        "description": "Posted",
                        "schema": {
                            "$ref": "#/definitions/dto.MerchantWithdrawResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "409": {
                        "description": "Concurrency conflict (retryable)",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Business rule violation",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
...
        "/v1/finance/topup": {
            "post": {
                "description": "Internal credit to a wallet; settles immediately. Idempotent on `reference` — a duplicate returns 200 with the original result.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Top up a wallet",
                "parameters": [
                    {
                        "description": "Top-up request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TopupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Duplicate idempotent replay",
                        "schema": {
                            "$ref": "#/definitions/dto.TopupResponse"
                        }
                    },
                    "201": {
                        "description": "Posted",
                        "schema": {
                            "$ref": "#/definitions/dto.TopupResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Business rule violation",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
...
        "/v1/finance/transfer": {
            "post": {
                "description": "In-book transfer (settles immediately, fee/VAT applied). Idempotent on `reference` — a duplicate returns 200.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Transfer between wallets",
                "parameters": [
                    {
                        "description": "Transfer request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Duplicate idempotent replay",
                        "schema": {
                            "$ref": "#/definitions/dto.TransferResponse"
                        }
                    },
                    "201": {
                        "description": "Posted",
                        "schema": {
                            "$ref": "#/definitions/dto.TransferResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "409": {
                        "description": "Concurrency conflict (retryable)",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Business rule violation (e.g. insufficient funds)",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/v1/finance/withdraw": {
            "post": {
                "description": "Debit a wallet for external disbursement. Ledger commits immediately; disbursement is then driven by the Treasury state machine. Idempotent on `reference`.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Withdraw from a wallet",
                "parameters": [
                    {
                        "description": "Withdraw request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.WithdrawRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Duplicate idempotent replay",
                        "schema": {
                            "$ref": "#/definitions/dto.WithdrawResponse"
                        }
                    },
                    "201": {
                        "description": "Posted",
                        "schema": {
                            "$ref": "#/definitions/dto.WithdrawResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "409": {
                        "description": "Concurrency conflict (retryable)",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Business rule violation (e.g. insufficient funds)",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
...
        "/v1/treasury/withdrawals/{ext_payout_ref}/acked": {
            "post": {
                "description": "S2S callback — Treasury accepted the payout into a batch.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "treasury"
                ],
                "summary": "Treasury: mark withdrawal acknowledged",
                "parameters": [
                    {
                        "type": "string",
                        "description": "External payout reference",
                        "name": "ext_payout_ref",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ack payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MarkResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Unknown payout reference",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Illegal state transition",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/v1/treasury/withdrawals/{ext_payout_ref}/completed": {
            "post": {
                "description": "S2S callback — payout settled at the rail (NAPAS ref captured).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "treasury"
                ],
                "summary": "Treasury: mark withdrawal completed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "External payout reference",
                        "name": "ext_payout_ref",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Completion payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CompletedRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MarkResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Unknown payout reference",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Illegal state transition",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/v1/treasury/withdrawals/{ext_payout_ref}/disbursing": {
            "post": {
                "description": "S2S callback — Treasury has submitted the payout to the rail.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "treasury"
                ],
                "summary": "Treasury: mark withdrawal disbursing",
                "parameters": [
                    {
                        "type": "string",
                        "description": "External payout reference",
                        "name": "ext_payout_ref",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MarkResponse"
                        }
                    },
                    "404": {
                        "description": "Unknown payout reference",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Illegal state transition",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/v1/treasury/withdrawals/{ext_payout_ref}/reverse": {
            "post": {
                "description": "S2S callback — payout failed/timed out; reverse the ledger debit. Idempotent (re-reversal returns was_already_reversed=true).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "treasury"
                ],
                "summary": "Treasury: reverse a failed withdrawal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "External payout reference",
                        "name": "ext_payout_ref",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reversal payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ReversalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ReversalResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Unknown payout reference",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Illegal state transition",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.MerchantWithdrawRequest": {
            "type": "object",
            "required": [
                "amount",
                "group_id",
                "reference"
            ],
            "properties": {
                "amount": {
                    "type": "string"
                },
                "auto_sweep": {
                    "description": "AutoSweep defaults to true (pointer distinguishes \"omitted\" from \"false\").",
                    "type": "boolean"
                },
                "ext_payout_ref": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 8
                },
                "group_id": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2
                },
                "reference": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 8
                }
            }
        },
        "dto.MerchantWithdrawResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "event_uuid": {
                    "type": "string"
                },
                "fee_gross": {
                    "type": "string"
                },
                "settlement_balance_after": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_deducted": {
                    "type": "string"
                },
                "tran_internal_key": {
                    "type": "integer"
                },
                "transaction_status": {
                    "description": "ISO 20022 (§13.3)",
                    "type": "string"
                },
                "vat_amount": {
                    "type": "string"
                }
            }
        },
...
        "dto.TopupRequest": {
            "type": "object",
            "required": [
                "acct_no",
                "amount",
                "reference"
            ],
            "properties": {
                "acct_no": {
                    "type": "string"
                },
                "amount": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "narrative": {
                    "type": "string",
                    "maxLength": 250
                },
                "reference": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 8
                }
            }
        },
...
    },
...
}